+++ CHANGELOG.md
@@ -1,3 +1,19 @@
+## v0.3.7 (2018-05-22)
+
+- chore(web): update `rocket` to `v0.3.11`
+- chore(*): update dependencies
+
+## v0.3.6 (2018-05-13)
+
+- new(csv-export): export average rating of an entry
+- chore(web): update `rocket` to `v0.3.10`
+- chore(*): update dependencies
+
+## v0.3.5 (2018-04-27)
+
+- new(csv-export): export category names instead of IDs
+- chore(web): update `rocket` to `v0.3.9`
+
 ## v0.3.4 (2018-04-24)
 
 - new(csv-export): implement csv-export of entries for a given bbox